refactor(store): consolidate duplicate auth types and remove dead code
- Remove unused UpdateSuccessPayload type from types.ts (dead code)
- Move AuthState and LoginPayload from authSlice.ts to types.ts
- Replace inline PersistedAuthState in store/index.ts with shared AuthState
- Remove inline interface definitions from authSlice.ts

Now auth types have a single source of truth in types.ts,
preventing type drift between authSlice.ts and store/index.ts.
