+++ src/store/authSlice.ts
@@ -1,22 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-
-export interface AuthState {
-  isAuthenticated: boolean;
-  user: {
-    id: number;
-    username: string;
-    image?: string;
-    is_staff: boolean;
-    is_superuser: boolean;
-    logins_remaining_for_staff: number;
-    staff_access_granted: boolean;
-    active_role: 'regular' | 'staff' | 'superuser';
-    role_label: string;
-  } | null;
-  accessToken: string | null;
-  refreshToken: string | null;
-  showLogoutMessage: boolean;
-}
+import { AuthState, LoginPayload } from "./types";
 
 const initialState: AuthState = {
   isAuthenticated: false,
@@ -25,19 +8,6 @@ const initialState: AuthState = {
   refreshToken: null,
   showLogoutMessage: false,
 };
-
-interface LoginPayload {
-  id: number;
-  username: string;
-  access: string;
-  refresh: string;
-  is_staff: boolean;
-  is_superuser: boolean;
-  logins_remaining_for_staff: number;
-  staff_access_granted: boolean;
-  active_role: 'regular' | 'staff' | 'superuser';
-  role_label: string;
-}
 
 const authSlice = createSlice({
   name: "auth",
